closure-compiler middleware improved, decides between the raw compiler or the calcdeps.py script based on the data in .deployerinfo.json
If there is an "input" file specified in the compilation target sheet, the calcdeps.py script is used. It then looks for dependencies in all files from directories specified in an optional "paths" array. If there is a "$GOOG" in this array, the closure/goog directory is automatically injected in the array.